+++ backend/orchestrator/computeplan_pb2_grpc.py
@@ -34,6 +34,11 @@ class ComputePlanServiceStub(object):
                 request_serializer=computeplan__pb2.QueryPlansParam.SerializeToString,
                 response_deserializer=computeplan__pb2.QueryPlansResponse.FromString,
                 )
+        self.UpdatePlan = channel.unary_unary(
+                '/orchestrator.ComputePlanService/UpdatePlan',
+                request_serializer=computeplan__pb2.UpdateComputePlanParam.SerializeToString,
+                response_deserializer=computeplan__pb2.UpdateComputePlanResponse.FromString,
+                )
 
 
 class ComputePlanServiceServicer(object):
@@ -63,6 +68,12 @@ class ComputePlanServiceServicer(object):
         context.set_details('Method not implemented!')
         raise NotImplementedError('Method not implemented!')
 
+    def UpdatePlan(self, request, context):
+        """Missing associated documentation comment in .proto file."""
+        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
+        context.set_details('Method not implemented!')
+        raise NotImplementedError('Method not implemented!')
+
 
 def add_ComputePlanServiceServicer_to_server(servicer, server):
     rpc_method_handlers = {
@@ -85,6 +96,11 @@ def add_ComputePlanServiceServicer_to_server(servicer, server):
                     servicer.QueryPlans,
                     request_deserializer=computeplan__pb2.QueryPlansParam.FromString,
                     response_serializer=computeplan__pb2.QueryPlansResponse.SerializeToString,
+            ),
+            'UpdatePlan': grpc.unary_unary_rpc_method_handler(
+                    servicer.UpdatePlan,
+                    request_deserializer=computeplan__pb2.UpdateComputePlanParam.FromString,
+                    response_serializer=computeplan__pb2.UpdateComputePlanResponse.SerializeToString,
             ),
     }
     generic_handler = grpc.method_handlers_generic_handler(
@@ -161,5 +177,22 @@ class ComputePlanService(object):
         return grpc.experimental.unary_unary(request, target, '/orchestrator.ComputePlanService/QueryPlans',
             computeplan__pb2.QueryPlansParam.SerializeToString,
             computeplan__pb2.QueryPlansResponse.FromString,
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
+
+    @staticmethod
+    def UpdatePlan(request,
+            target,
+            options=(),
+            channel_credentials=None,
+            call_credentials=None,
+            insecure=False,
+            compression=None,
+            wait_for_ready=None,
+            timeout=None,
+            metadata=None):
+        return grpc.experimental.unary_unary(request, target, '/orchestrator.ComputePlanService/UpdatePlan',
+            computeplan__pb2.UpdateComputePlanParam.SerializeToString,
+            computeplan__pb2.UpdateComputePlanResponse.FromString,
             options, channel_credentials,
             insecure, call_credentials, compression, wait_for_ready, timeout, metadata)