+++ backend/orchestrator/datamanager_pb2_grpc.py
@@ -29,6 +29,11 @@ class DataManagerServiceStub(object):
                 request_serializer=datamanager__pb2.QueryDataManagersParam.SerializeToString,
                 response_deserializer=datamanager__pb2.QueryDataManagersResponse.FromString,
                 )
+        self.UpdateDataManager = channel.unary_unary(
+                '/orchestrator.DataManagerService/UpdateDataManager',
+                request_serializer=datamanager__pb2.UpdateDataManagerParam.SerializeToString,
+                response_deserializer=datamanager__pb2.UpdateDataManagerResponse.FromString,
+                )
 
 
 class DataManagerServiceServicer(object):
@@ -52,6 +57,12 @@ class DataManagerServiceServicer(object):
         context.set_details('Method not implemented!')
         raise NotImplementedError('Method not implemented!')
 
+    def UpdateDataManager(self, request, context):
+        """Missing associated documentation comment in .proto file."""
+        context.set_code(grpc.StatusCode.UNIMPLEMENTED)
+        context.set_details('Method not implemented!')
+        raise NotImplementedError('Method not implemented!')
+
 
 def add_DataManagerServiceServicer_to_server(servicer, server):
     rpc_method_handlers = {
@@ -69,6 +80,11 @@ def add_DataManagerServiceServicer_to_server(servicer, server):
                     servicer.QueryDataManagers,
                     request_deserializer=datamanager__pb2.QueryDataManagersParam.FromString,
                     response_serializer=datamanager__pb2.QueryDataManagersResponse.SerializeToString,
+            ),
+            'UpdateDataManager': grpc.unary_unary_rpc_method_handler(
+                    servicer.UpdateDataManager,
+                    request_deserializer=datamanager__pb2.UpdateDataManagerParam.FromString,
+                    response_serializer=datamanager__pb2.UpdateDataManagerResponse.SerializeToString,
             ),
     }
     generic_handler = grpc.method_handlers_generic_handler(
@@ -128,5 +144,22 @@ class DataManagerService(object):
         return grpc.experimental.unary_unary(request, target, '/orchestrator.DataManagerService/QueryDataManagers',
             datamanager__pb2.QueryDataManagersParam.SerializeToString,
             datamanager__pb2.QueryDataManagersResponse.FromString,
+            options, channel_credentials,
+            insecure, call_credentials, compression, wait_for_ready, timeout, metadata)
+
+    @staticmethod
+    def UpdateDataManager(request,
+            target,
+            options=(),
+            channel_credentials=None,
+            call_credentials=None,
+            insecure=False,
+            compression=None,
+            wait_for_ready=None,
+            timeout=None,
+            metadata=None):
+        return grpc.experimental.unary_unary(request, target, '/orchestrator.DataManagerService/UpdateDataManager',
+            datamanager__pb2.UpdateDataManagerParam.SerializeToString,
+            datamanager__pb2.UpdateDataManagerResponse.FromString,
             options, channel_credentials,
             insecure, call_credentials, compression, wait_for_ready, timeout, metadata)