+++ backend/substrapp/tests/views/test_views_datamanager.py
@@ -159,6 +159,28 @@ class DataManagerViewTests(APITestCase):
         shutil.rmtree(MEDIA_ROOT, ignore_errors=True)
         self.logger.setLevel(self.previous_level)
 
+    def test_datamanager_update(self):
+        data_manager = self.expected_results[0]
+        data = {
+            "key": data_manager["key"],
+            "name": "Bar",
+        }
+
+        url = reverse("substrapp:data_manager-detail", args=[data_manager["key"]])
+        data_manager["name"] = data["name"]
+
+        with mock.patch.object(OrchestratorClient, "update_datamanager", side_effect=data_manager):
+            response = self.client.put(url, data=data, format="json", **self.extra)
+
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        error = OrcError()
+        error.code = StatusCode.INTERNAL
+
+        with mock.patch.object(OrchestratorClient, "update_datamanager", side_effect=error):
+            response = self.client.put(url, data=data, format="json", **self.extra)
+        self.assertEqual(response.status_code, status.HTTP_500_INTERNAL_SERVER_ERROR)
+
     def test_datamanager_list_empty(self):
         DataManagerRep.objects.all().delete()
         response = self.client.get(self.url, **self.extra)